+++ tests/testthat/test-archive_extract.R
@@ -61,5 +61,23 @@ describe("archive_extract", {
 
     expect_equal(files, filename)
   })
+
+  it("can strip components if desired", {
+    in_dir <- tempfile()
+    out_dir <- tempfile()
+    on.exit(unlink(c(in_dir, out_dir), recursive = TRUE))
+
+    dir.create(file.path(in_dir, "foo/bar"), recursive = TRUE)
+
+    write.csv(iris, file.path(in_dir, "foo", "bar", "iris.csv"))
+    write.csv(mtcars, file.path(in_dir, "foo", "mtcars.csv"))
+
+    ar <- tempfile(fileext = ".zip")
+    archive_write_dir(ar, in_dir)
+
+    archive_extract(ar, out_dir, strip_components = 1)
+
+    expect_true(all(c("bar/iris.csv", "mtcars.csv") %in% list.files(out_dir, recursive = TRUE)))
+  })
 })