+++ NEWS.md
@@ -1,5 +1,7 @@
 # archive (development version)
 
+* `archive_extract()` gains a `strip_components` argument to strip leading pathname components if desired (#27)
+
 * `archive()`, `archive_read()`, `archive_extract()` and `file_read()` now all accept R connections in addition to file paths.
   This allows you to do things like read remote archives with a `url()` or `curl::curl()` connection like you would a file on disk (#7)